[#8] feat: 반영 리뷰 (상수 분리, semantic 태그 적용)
